SpecialOffer & Run Entities - Doctrine DateTime Fix
When Doctrine saves a DateTime object to the database it does not check or account for its time zone instead simply formatting it as a string and saving it.  When that DateTime is later hydrated from the database the assumption is made that the object was in PHP's default time zone.  This is clearly problematic where sane handling of dates and times is expected (i.e. persist UTC times and hydrate with a UTC time zone).

Added the Utility\DateTime class which contains the static methods toDoctrine and fromDoctrine which are integrated inte the SpecialOffer and Run entities to compensate for Doctrine's deficient DateTime handling.  The integration of these methods insures that DateTime objects are saved as a UTC timestamp and hydrated with the proper time and a UTC time zone.