64 bit PHP
Due to the fact Shopify IDs are very large integers (I've observed ~30 billion as a variant ID and ~950 billion as a product ID) 64 bit PHP is required to deal with them.  Added a README.md with this made explicit.  Also added checks wherever variant IDs are dealt with which throws a LogicException if the version of PHP is not at least 64 bits.