Basic Homepage
Displays a very basic listing of pending and active offers and adds a button to the Shopify bar (above the iframe in which the app is displayed) to create a new offer.

All links (including the create button) are currently "broken": The controller actions invoked by them simply throw a \LogicException indicating that they are unimplemented.