ToHttpsListener & Scheme Fix
When working behind ngrok (in development) I was relying on setting the scheme of routes to "https" to generate the appropriate links (despite the fact the actual link to the server was HTTP).  Unfortunately this caused Symfony to continually reject incoming requests and attempt to redirect to HTTPS (as the link between ngrok and Symfony was HTTP) leading to a redirect loop.

The solution was to implement an event listener which intercepts all incoming requests and sets the router's RequestContext object's scheme to "https" so that the router believes all requests to be HTTPS and generates HTTPS URLs without issue.