LinkedDateTimeType
Extends Symfony's DateTimeType and allows two date/time pickers to be "linked": I.e. one defines a beginning and the other an end.  This link prevents the end from being set before the beginning, or the beginning being set after the end.

Does not currently implement this logic server side.  I'm not sure if this is even possible within the Symfony form framework.