EmailNotificationListener - Templating Support
Under the "notifications" key in the configuration "start_template" and "end_template" may now be provided giving the names of Twig templates which will be used for the bodies of the start and end emails, respectively.

The EmailNotificationListener automatically detects (based on file extension) whether these templates produce HTML or plaintext and sets the content type of the resulting email accordingly.